Create a page helper.js file, that deals with interactions with the page. And add comments to clarify code so its immediately clear what code is doing and you do not want to take the time to read it
