+++ lib/end-dash.js
@@ -3,6 +3,7 @@ var _ = require('underscore'),
     TemplateStore = require('./template_store'),
     AttributeReaction = require('./reactions/attribute'),
     CollectionReaction = require('./reactions/collection'),
+    PageHelper = require('./page_helper'),
     ModelReaction = require('./reactions/model'),
     VariableReaction = require('./reactions/variable'),
     ConditionalReaction = require('./reactions/conditional'),
@@ -10,6 +11,7 @@ var _ = require('underscore'),
     ViewReaction = require('./reactions/view'),
     ScopeReaction = require('./reactions/scope'),
     DebuggerReaction = require('./reactions/debugger');
+
 
   Parser.registerReaction(PartialReaction);
   Parser.registerReaction(ScopeReaction);
@@ -34,15 +36,15 @@ EndDash.registerTemplate = function(templatePath, markup) {
 
 EndDash.getTemplateClass = EndDash.getTemplate = function(templatePath){
   return TemplateStore.getTemplate(templatePath)
-}
+} //getTemplate is deprecated but still referenced in some outside systems
 
 EndDash.clearAndReload = function() {
   TemplateStore.clear()
-  this.loadPage()
+  this.loadTemplatesFromPage()
 }
 
-EndDash.loadFromPage = function() {
-  TemplateStore.lazyLoadFromPage()
+EndDash.loadTemplatesFromPage = function() {
+  PageHelper.lazyLoadFromPage()
 }
 
 EndDash.isTemplateLoaded = function(templatePath) { //Here for testing
