+++ lib/template.js
@@ -19,7 +19,7 @@ function Template(model, opts) {
   if(!this.markup) {
     throw new Error("Created template without markup")
   }
-  this.el = this.template = this.markup.clone()
+  this.el = this.template = this.markup.clone() // template is deprecated but used in some outside systems
   this.bind()
 }
