+++ lib/template_store.js
@@ -26,18 +26,6 @@ var _parseTemplate = function(templatePath) {
   })).generate();
 };
 
-var getTemplatesOnPage = function() {
-  var scriptTags = window.document.getElementsByTagName("SCRIPT")
-  return _.filter(scriptTags, function(scriptElem) {
-    return scriptElem.type == "EndDash"
-  })
-}
-
-//Trims leading and ending whitespace, necessary for JQuery create
-var normalizeTemplate = function(template) {
-  return template.textContent.replace(/^\s*/, "").replace(/\s*$/, "")
-}
-
 TemplateStore.isLoaded = function(templatePath) {
   return !!RAW_TEMPLATES[templatePath];
 };
@@ -70,17 +58,6 @@ TemplateStore.getTemplate = function(templatePath) {
   return TEMPLATES[name];
 };
 
-//Lazy loading saves time since there are many templates per page and few used
-//Important, order of loading HTML and parsing matters. Those templates with
-//partials depend on their partials being loaded as raw HTML before they go
-//Through the parser
-TemplateStore.lazyLoadFromPage = function (){
-  var that = this
-  _(getTemplatesOnPage()).each(function(template){
-    that.load(template.getAttribute('name'), normalizeTemplate(template))
-  })
-}
-
 TemplateStore.clear = function() {
     RAW_TEMPLATES = {}
     TEMPLATES = {}
@@ -91,7 +68,6 @@ _.bindAll(TemplateStore, 'isLoaded',
                          'load',
                          'loadAndParse',
                          'getTemplate',
-                         'lazyLoadFromPage',
                          'clear');
 
 module.exports = TemplateStore;
